+++ src/Pyro4/nsc.py
@@ -71,7 +71,6 @@ def handleCommand(nameserver, options, args):
 
 def main(args):
     from optparse import OptionParser
-
     usage = "usage: %prog [options] command [arguments]\nCommand is one of: " \
             "register remove removematching list listmatching ping"
     parser = OptionParser(usage=usage)